Fixing removal of $ from config values
This fixes the remove of dollar signs in config values (for example passwords). Before the install would fail if your password had a $ in it because what was returned did not match what was passed in (Example: If your password is Pa$sw0rd it would write Pasw0rd to the config, thus triggering the error on line 52.